1.3.9
v1.3.9 Removed main button. Open/close legend via custom key press "N"(default for now) changeable in settings.
*Legend now on gui.left - recommend using Mooncats Vertical left gui mod till 0.15 is released
