+++ info.json
@@ -1,7 +1,7 @@
 {
 	"name":"Enhanced_Map_Colors",
 	"author":"osldgoth",
-	"version":"1.3.70",
+	"version":"1.3.9",
 	"title":"Enhanced Map Colors",
 	"homepage":"https://github.com/osldgoth/Enhanced_Map_Colors/releases",
 	"description":"Adds color to items on map/minimap for contrast. Aids map readability.",
